chore: Rename package to 'bayesian_pyhf'
* Rename files to 'bayesian_pyhf' paths.
* Change every instance of `Bayesian_pyhf` to `bayesian_pyhf` using

```
git grep --name-only "Bayesian\_pyhf" | xargs sed -i 's/Bayesian\_pyhf/bayesian\_pyhf/g'
```
