+++ README.md
@@ -2,7 +2,7 @@
 The aim of this project is to provide tools for Bayesian analysis of ``pyhf`` models using the Python library ``PyMC``.
 
 ## Setup
-To set up a Python virtual environment, download the following dependency file [environment.yml](https://github.com/malin-horstmann/Bayesian_pyhf/blob/main/environment.yml) and run:
+To set up a Python virtual environment, download the following dependency file [environment.yml](https://github.com/malin-horstmann/bayesian_pyhf/blob/main/environment.yml) and run:
 
 ```
 conda env create --file environment.yml
@@ -12,9 +12,9 @@ You can then run
 ```
 pip install --editable .
 ```
-for an editable install of `Bayesian_pyhf`.
+for an editable install of `bayesian_pyhf`.
 
-For a quick example, see [example.ipynb](https://github.com/malin-horstmann/Bayesian_pyhf/blob/main/examples/SimpleEx.ipynb).
+For a quick example, see [example.ipynb](https://github.com/malin-horstmann/bayesian_pyhf/blob/main/examples/SimpleEx.ipynb).
 
 ## Development
 
@@ -35,7 +35,7 @@ nox
 An environment can then be created from the lock file either with
 
 ```
-conda-lock install --name Bayesian_pyhf conda-lock.yml
+conda-lock install --name bayesian_pyhf conda-lock.yml
 ```
 
 or
@@ -52,5 +52,5 @@ To add new dependencies to the environment definition files simply add the depen
 To update your existing environment from an updated `environment.yml` file use
 
 ```
-conda env update --name Bayesian_pyhf --file environment.yml
+conda env update --name bayesian_pyhf --file environment.yml
 ```
